bug: assembly 启动失败[1.5.4] #4475 消除重复命名
